added actions, reducers
also changed names of 'Calculator' to 'App' and 'Board' to 'Calculator' for clarification
